[refactor] Adapt `server` for new structure
